refactored the core/auth/viewsets & core/auth/serializers. Now ready for using these as base templates
