+++ core/auth/serializers/__init__.py
@@ -1,0 +1,3 @@
+# core/auth/serializers/__init__.py 
+from .register import RegisterSerializer
+from .login import LoginSerializer
