+++ core/auth/serializers/register.py
@@ -1,30 +1,10 @@
-# core/auth/serializers.py
+# core/auth/serializer/register.py
 
 from rest_framework import serializers
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-from rest_framework_simplejwt.settings import api_settings
-from django.contrib.auth.models import update_last_login
 from django.core.exceptions import ObjectDoesNotExist
 
 from core.user.serializers import UserSerializer
 from core.user.models import User
-
-
-class LoginSerializer(TokenObtainPairSerializer):
-    
-    def validate(self, attrs):
-        data = super().validate(attrs)
-
-        refresh = self.get_token(self.user)
-
-        data['user'] = UserSerializer(self.user).data
-        data['refresh'] = str(refresh)
-        data['access'] = str(refresh.access_token)
-
-        if api_settings.UPDATE_LAST_LOGIN:
-            update_last_login(None, self.user)
-
-        return data
 
 
 class RegisterSerializer(UserSerializer):
