+++ core/auth/viewsets/__init__.py
@@ -1,0 +1,5 @@
+# core/auth/viewsets/__init__.py
+
+from .register import RegistrationViewSet
+from .login import LoginViewSet
+from .refresh import RefreshViewSet
