+++ core/auth/viewsets/login.py
@@ -1,0 +1,25 @@
+# core/auth/viewsets/login.py
+
+from rest_framework.response import Response
+from rest_framework_simplejwt.views import TokenObtainPairView
+from rest_framework.viewsets import ModelViewSet
+from rest_framework.permissions import AllowAny
+from rest_framework import status
+from rest_framework_simplejwt.exceptions import TokenError, InvalidToken
+from core.auth.serializers import LoginSerializer
+
+
+class LoginViewSet(ModelViewSet, TokenObtainPairView):
+    serializer_class = LoginSerializer
+    permission_classes = (AllowAny,)
+    http_method_names = ['post']
+
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+
+        try:
+            serializer.is_valid(raise_exception=True)
+        except TokenError as e:
+            raise InvalidToken(e.args[0])
+
+        return Response(serializer.validated_data, status=status.HTTP_200_OK)
