+++ core/auth/viewsets/refresh.py
@@ -1,0 +1,23 @@
+# core/auth/viewsets/refresh.py
+
+from rest_framework.response import Response
+from rest_framework_simplejwt.views import TokenRefreshView
+from rest_framework.permissions import AllowAny
+from rest_framework import status
+from rest_framework import viewsets
+from rest_framework_simplejwt.exceptions import TokenError, InvalidToken
+
+
+class RefreshViewSet(viewsets.ViewSet, TokenRefreshView):
+    permission_classes = (AllowAny,)
+    http_method_names = ['post']
+
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+
+        try:
+            serializer.is_valid(raise_exception=True)
+        except TokenError as e:
+            raise InvalidToken(e.args[0])
+
+        return Response(serializer.validated_data, status=status.HTTP_200_OK)
